Move entropy and symmetric_key to public headers
