+++ include/ccf/crypto/entropy.h
