+++ include/ccf/crypto/symmetric_key.h
