+++ src/apps/js_generic/js_generic_base.cpp
@@ -6,7 +6,6 @@
 #include "ccf/historical_queries_adapter.h"
 #include "ccf/user_frontend.h"
 #include "ccf/version.h"
-#include "crypto/entropy.h"
 #include "crypto/key_wrap.h"
 #include "js/wrap.h"
 #include "kv/untyped_map.h"
