+++ src/crypto/test/crypto.cpp
@@ -4,9 +4,10 @@
 #include "ccf/crypto/base64.h"
 #include "ccf/crypto/key_pair.h"
 #include "ccf/crypto/rsa_key_pair.h"
+#include "ccf/crypto/symmetric_key.h"
 #include "crypto/certs.h"
 #include "crypto/csr.h"
-#include "crypto/entropy.h"
+#include "ccf/crypto/entropy.h"
 #include "crypto/hmac.h"
 #include "crypto/key_wrap.h"
 #include "crypto/openssl/key_pair.h"
@@ -14,7 +15,6 @@
 #include "crypto/openssl/symmetric_key.h"
 #include "crypto/openssl/verifier.h"
 #include "crypto/openssl/x509_time.h"
-#include "crypto/symmetric_key.h"
 #include "crypto/verifier.h"
 
 #include <chrono>
