+++ src/js/crypto.cpp
@@ -1,10 +1,10 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the Apache 2.0 License.
 
+#include "ccf/crypto/entropy.h"
 #include "ccf/crypto/rsa_key_pair.h"
 #include "ccf/crypto/sha256.h"
 #include "crypto/ecdsa.h"
-#include "crypto/entropy.h"
 #include "crypto/key_wrap.h"
 #include "js/wrap.h"
 #include "tls/ca.h"
