+++ src/node/share_manager.h
@@ -2,10 +2,10 @@
 // Licensed under the Apache 2.0 License.
 #pragma once
 
+#include "ccf/crypto/entropy.h"
 #include "ccf/crypto/rsa_key_pair.h"
+#include "ccf/crypto/symmetric_key.h"
 #include "ccf/ds/logger.h"
-#include "crypto/entropy.h"
-#include "crypto/symmetric_key.h"
 #include "kv/encryptor.h"
 #include "ledger_secrets.h"
 #include "network_state.h"
