+++ src/tls/context.h
@@ -5,7 +5,6 @@
 #include "ccf/crypto/base64.h"
 #include "ccf/ds/logger.h"
 #include "cert.h"
-#include "crypto/entropy.h"
 #include "tls/tls.h"
 
 #include <memory>
